+++ tests/testthat/test-Connection.R
@@ -7,3 +7,17 @@ test_that("warn about case-insensitve arguments", {
     check_args(list(xa = 1, xA = 1, xb = 1, xB = 1))
   })
 })
+
+test_that("build_connection_string handles simple inputs", {
+  expect_equal(build_connection_string(list()), "")
+  expect_equal(build_connection_string(list(foo = "1")), "foo=1")
+  expect_equal(build_connection_string(list(foo = "1", bar = "2")), "foo=1;bar=2")
+})
+
+test_that("build_connection_string automatically escapes if needed", {
+  expect_equal(build_connection_string(list(foo = "*")), "foo={*}")
+  # Already wrapped
+  expect_equal(build_connection_string(list(foo = "{*}")), "foo={*}")
+  # Respects AsIs
+  expect_equal(build_connection_string(list(foo = I("*"))), "foo=*")
+})